backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = "^3.11,<4.0"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
llama-index = "^0.11.2"
cachetools = "^5.3.3"
chromadb = "^0.5.12"
llama-index-vector-stores-chroma = "^0.2.0"
llama-index-callbacks-langfuse = "^0.2.0"
chromedriver-autoinstaller = "^0.6.4"



[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"



[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"